Handle floating-point rope theta values
Summary:
- Parse rope_theta from JSON as a floating-point value before storing it.

Details:
- Some model configs encode rope_theta as a JSON number with floating-point type.
- Read the value as double first to avoid type errors during transformer setup.

+++ models/transformer.cpp
@@ -129,7 +129,7 @@ void Transformer::setupParameters(json &cfg, json &generation_cfg,
                              ? cfg["sliding_window_pattern"].get<unsigned int>()
                              : 1;
   MAX_POSITION_EMBEDDINGS = cfg["max_position_embeddings"].get<unsigned int>();
-  ROPE_THETA = cfg["rope_theta"].get<unsigned int>();
+  ROPE_THETA = static_cast<unsigned int>(cfg["rope_theta"].get<double>());
   TIE_WORD_EMBEDDINGS = cfg["tie_word_embeddings"].get<bool>();
   NORM_EPS = cfg["rms_norm_eps"];
   GQA_SIZE = NUM_HEADS / NUM_KEY_VALUE_HEADS;
